add clientoption to send a csrf header for api calls, and triggering a login if authentication is required
the authState holds the csrf token, if any. api's can use that to check with a
cookie (where the two values should be present, known and related, not the
same).

if an api call responds with "user:noAuth" or "user:badAuth", a login can be
triggered. at most one login concurrently, other api calls that failed will
wait for the login. once the login has completed, the api calls are made again.
this is transparent to the call site.